move the recomm API tests to respective dirs of remote and local
