+++ tests/scripts/local_monitoring_tests/local_monitoring_tests.sh
@@ -43,7 +43,7 @@ function local_monitoring_tests() {
 	target="crc"
 	metric_profile_json="${METRIC_PROFILE_DIR}/resource_optimization_local_monitoring.json"
 
-	local_monitoring_tests=("sanity" "extended" "negative" "test_e2e" "test_e2e_pr_check" "test_bulk_api_ros" "runtimes")
+	local_monitoring_tests=("sanity" "extended" "negative" "test_e2e" "test_e2e_pr_check" "test_bulk_api_ros" "runtimes" "recommendation")
 
 	# check if the test case is supported
 	if [ ! -z "${testcase}" ]; then
