+++ tests/scripts/local_monitoring_tests/pytest.ini
@@ -9,3 +9,4 @@ markers =
     test_bulk_api_ros: mark test as a bulk_api ros test
     layers: mark test as a layer-related test (createLayer, listLayers)
     runtimes: mark test as a runtimes test
+    recommendation: mark test as a recommendation API test
