--- tests/scripts/recommendation_tests/pytest.ini
@@ -1,5 +1,0 @@
-# content of pytest.ini
-[pytest]
-markers =
-    remote: mark test as an RM test
-    local: mark test as an LM test
