--- tests/scripts/recommendation_tests/requirements.txt
@@ -1,6 +1,0 @@
-pytest
-requests
-jinja2
-pytest-html==3.2.0
-kubernetes
-docker
