+++ tests/scripts/remote_monitoring_tests/pytest.ini
@@ -6,3 +6,4 @@ markers =
     negative: mark test as a negative test
     extended: mark test as a extended test
     perf_profile: mark test as a performance profile test
+    recommendation: mark test as a recommendation API test
